refactor(volatility-forecast): clean up typo file, add missing exports, remove empty files
- Remove duplicate model_managemer.py (typo); model_manager.py already exists
- Export OffsetLogReturnTemplate, SignatureFeaturesTemplate, FadingMemorySignatureTemplate from features/__init__.py
- Remove 3 accidental empty files from repo root (addtime-, basepoint-, leadlag)

+++ volatility_forecast/model/volgru_torch.py
@@ -19,7 +19,7 @@ class VolGRUCellTorch(nn.Module):
         config: VolGRUConfig,
         n_features: int,
         *,
-        dtype: torch.dtype = torch.float64,
+        dtype: torch.dtype = torch.float32,
     ) -> None:
         super().__init__()
 
@@ -205,7 +205,7 @@ class VolGRUModelTorch(nn.Module):
         config: VolGRUConfig,
         n_features: int,
         *,
-        dtype: torch.dtype = torch.float64,
+        dtype: torch.dtype = torch.float32,
     ) -> None:
         super().__init__()
         self.config = config
